Home page UI fix & simultaneous HTTP crash fix
Fixed a bug where after connecting with an account and move to the Home page, the TextView message does not change as the authentication is happening.
Also fixed a bug where going into Home page multiple times through the hamburger menu calls the same HTTP post request multiple times that caused the app to crash.